internal: change features flag from int to uin64t_t

+++ libnodegl/gctx.h
@@ -105,7 +105,7 @@ struct gctx {
     struct ngl_config config;
     const struct gctx_class *class;
     int version;
-    int features;
+    uint64_t features;
     struct limits limits;
 };
